Crate wgpu_glyph

A fast text renderer for wgpu. Powered by glyph_brush.

Modules

ab_glyph

Re-exported ab_glyph types.

Structs

Extra

Default extra field type. Non-layout data for vertex generation.

FontId

Id for a font.

GlyphBrush

Object allowing glyph drawing, containing cache state. Manages glyph positioning cacheing, glyph draw caching & efficient GPU texture cache updating and re-sizing on demand.

GlyphBrushBuilder

Builder for a GlyphBrush.

OwnedSection

OwnedText

Region

A region of the screen.

Section

An object that contains all the info to render a varied section of text. That is one including many parts with differing fonts/scales/colors bowing to a single layout.

SectionGeometry

SectionGlyph

A positioned glyph with info relating to the SectionText from which it was derived.

SectionText

Text to layout together using a font & scale.

Text

SectionText + extra.

Enums

BuiltInLineBreaker

Built-in linebreaking logic.

HorizontalAlign

Describes horizontal alignment preference for positioning & bounds.

Layout

Built-in GlyphPositioner implementations.

LineBreak

Indicator that a character is a line break, soft or hard. Includes the offset (byte-index) position.

VerticalAlign

Describes vertical alignment preference for positioning & bounds. Currently a placeholder for future functionality.

Traits

GlyphCruncher

Common glyph layout logic.

GlyphPositioner

Logic to calculate glyph positioning using Font, SectionGeometry and SectionText.

LineBreaker

Producer of a LineBreak iterator. Used to allow to the Layout to be line break aware in a generic way.

Functions

orthographic_projection

Helper function to generate a generate a transform matrix.

Type Definitions

SectionGlyphIter

SectionGlyph iterator.